requirements.txt
# Django 및 핵심 의존성
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0

# Python 실행 환경
RestrictedPython>=6.0.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
scipy>=1.10.0

# 비동기 작업 (선택)
celery>=5.3.0
redis>=4.5.0

# 보안
django-environ>=0.10.0

# 유틸리티
python-dateutil>=2.8.0

# 개발 도구
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0

# JupyterHub
jupyterhub>=4.0.0
jupyter>=1.0.0
notebook>=7.0.0
jupyterlab>=4.0.0
django-jupyterhub>=0.1.0  # Django 통합 (선택)

